feat: 사다리 높이를 Height로 포장
