+++ src/main/java/domain/Ladder.java
@@ -4,29 +4,17 @@ import util.LineItemGenerator;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.regex.Pattern;
 
 public class Ladder {
 
-    private static final Pattern NATURAL_NUMBER_FORMAT_REGEX = Pattern.compile("^[1-9][0-9]*$");
-
-    private final int height;
     private final List<Line> ladder;
 
-    public Ladder(String height) {
-        validateHeight(height);
-        this.height = Integer.parseInt(height);
+    public Ladder() {
         this.ladder = new ArrayList<>();
     }
 
-    private void validateHeight(String height) {
-        if (height == null || !NATURAL_NUMBER_FORMAT_REGEX.matcher(height).matches()) {
-            throw new IllegalArgumentException("사다리의 최대 높이는 자연수여야 합니다.");
-        }
-    }
-
-    public List<Line> makeLadder(int columnLength, LineItemGenerator lineItemGenerator) {
-        for (int i = 0; i < height; i++) {
+    public List<Line> makeLadder(Height height, int columnLength, LineItemGenerator lineItemGenerator) {
+        for (int i = 0; i < height.getHeight(); i++) {
             Line line = new Line(columnLength);
             line.makeLine(lineItemGenerator);
